Ripped HeartbeatCallback apart.
